Merge expac into src/pacman
This introduces code which has long been maintained at:

https://github.com/falconindy/expac

From the README:

expac is a data extraction tool for alpm databases. It features
printf-like flexibility and aims to be used as a simple tool for other
pacman based utilities which don't link against the library. It uses
pacman.conf as a config file for locating and loading your local and
sync databases.